feature.call_service: move packages to the correct core package
