+++ app/src/main/java/com/hcl/notflixpoc/core/domain/model/MovieDomainModel.kt
@@ -1,4 +1,4 @@
-package com.hcl.notflixpoc.domain.model
+package com.hcl.notflixpoc.core.domain.model
 
 data class MovieDomainModel(
     val name: String,
